nuttx/project-apps: do not suppress -Werror but suppress -Wshadow
This allows propagation of -Werror to the Meson projects but it removes
-Wshadow. It is common that variable shadowing is used but in NuttX the
code is required to not use shadowing. Thus we do not want to propagate
this flag to the Meson project. If projects also wants to enforce disuse
of the variable shadowing then it should add the flag on its own.